Installed AutoMapper library and mapped student model.

+++ ThirdWebAPI/Controllers/StudentController.cs
@@ -1,9 +1,11 @@
-﻿using Microsoft.AspNetCore.Mvc;
+﻿using AutoMapper;
+using Microsoft.AspNetCore.Mvc;
 using System.Collections;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 using ThirdWebAPI.Models;
 using ThirdWebAPI.Repositories;
+using WebAPICRUD.DTOs;
 
 namespace ThirdWebAPI.Controllers
 {
@@ -12,49 +14,62 @@ namespace ThirdWebAPI.Controllers
     public class StudentController : ControllerBase
     {
         private readonly IStudentRepository _repository;
+        private readonly IMapper _mapper;
 
-        public StudentController(IStudentRepository repository)
+        public StudentController(IStudentRepository repository,IMapper mapper)
         {
             this._repository = repository;
+            _mapper = mapper;
         }
 
         [HttpGet]
-        [ProducesResponseType(200, Type = typeof(IEnumerable<Student>))]
+        [ProducesResponseType(200, Type = typeof(IEnumerable<StudentReadDto>))]
         public IActionResult GetAllStudets()
         {
-            return Ok(_repository.GetStudents().GetAwaiter().GetResult());
+            var students = _repository.GetStudents().GetAwaiter().GetResult();
+            var mapStudents=_mapper.Map<IEnumerable<StudentReadDto>>(students);
+            return Ok(mapStudents);
         }
 
         [HttpGet("{id}")]
-        [ProducesResponseType(200, Type = typeof(Student))]
-        public IActionResult GetStudet(int id)
+        [ProducesResponseType(202, Type = typeof(StudentReadDto))]
+        public async Task<IActionResult> GetStudet(int id)
         {
-            var student = _repository.GetStudent(id).GetAwaiter().GetResult();
-            if(student == null)
+            var student = await _repository.GetStudent(id);
+            if (student == null)
                 return NotFound();
-            return Ok(student);
+            var studentRead = _mapper.Map<StudentReadDto>(student);
+            return Ok(studentRead);
         }
 
         [HttpPost]
-        [ProducesResponseType(200, Type = typeof(Student))]
-        public IActionResult CreateStudent(Student student)
+        [ProducesResponseType(200, Type = typeof(StudentCreateDto))]
+        public async Task<IActionResult> CreateStudent(StudentCreateDto student)
         {
-            _repository.CreateStudent(student);
-            return Ok(student);
+            var mapped = _mapper.Map<Student>(student);
+            await _repository.CreateStudent(mapped);
+            return Created("",student);
         }
 
-        [HttpDelete]
-        public IActionResult DeleteStudent(int id)
+        [HttpDelete("{id}")]
+        public async Task<IActionResult> DeleteStudent(int id)
         {
-            _repository.DeleteStudent(id);
+            var studentFromRepo = await _repository.GetStudent(id);
+            if (studentFromRepo == null)
+                return NotFound();
+            await _repository.DeleteStudent(id);
             return Ok("Deleted");
         }
 
         [HttpPut("{id}")]
-        public IActionResult UpdateStudent(int id, Student student)
+        public async Task<IActionResult> UpdateStudent(int id, StudentUpdateDto student)
         {
-           _repository.UpdateSTudent(id, student);
-            return Ok(student);
+            var studentFromRepo = await _repository.GetStudent(id);
+            if (studentFromRepo == null)
+                return NotFound();
+
+            _mapper.Map(student, studentFromRepo);
+            return Ok();
         }
     }
 }
